LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-async-fetch-mutation-clear-expected.txt


FAIL Clearing adoptedStyleSheets before fetch prevents styles from being applied. assert_equals: Before clearing: expected 1 placeholder(s). expected 1 but got 0
